misc/HtmlTestSuite/results/test-0297.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml">
<head>
<meta charset="utf-8"/>
<title>CSS basic box model: parsing float with invalid values</title>
<link href="mailto:ericwilligers@chromium.org" rel="author" title="Eric Willigers"/>
<link href="https://drafts.csswg.org/css-box-3/#propdef-float" rel="help"/>
<meta content="float supports only the grammar 'left | right | top | bottom | start | end | none | &lt;page-floats&gt;'." name="assert"/>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<script src="/css/support/parsing-testcommon.js"/>
</head>
<body>
<script>
test_invalid_value("float", "auto");
test_invalid_value("float", "right bottom");
test_invalid_value("float", "top, left");
</script>
</body>
</html>
